backend/agents/requirements.txt
# IELTS Reading Feedback Service - Dependencies

# LangChain and AI
langchain==0.3.13
langchain-openai==0.2.14
langchain-core==0.3.28
langchain-community==0.3.13

# OpenAI
openai>=1.58.1,<2.0.0

# FastAPI
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Vector Store - FAISS (исправлена версия!)
faiss-cpu==1.9.0.post1

# Utilities
python-dotenv==1.0.1
tiktoken==0.8.0
httpx==0.28.1
typing-extensions==4.12.2
python-multipart==0.0.20